+++ src/blocks/Table/models.ts
@@ -1,0 +1,5 @@
+export const TableMicrodataValues = {
+    TableType: 'https://schema.org/Table',
+    AboutProp: 'about',
+    accessModeProp: 'accessMode',
+} as const;