+++ src/components/Media/Video/Video.tsx
@@ -46,28 +46,10 @@ const Video = (props: VideoAllProps) => {
 
     const ref = React.useRef<HTMLVideoElement>(null);
 
+    React.useImperativeHandle(video.ref, () => ref.current, []);
+
     React.useEffect(() => {
         if (ref && ref.current) {
-            const {loop} = video;
-
-            if (loop && typeof loop !== 'boolean') {
-                const {start = 0, end} = loop;
-
-                ref.current.addEventListener(
-                    'timeupdate',
-                    () => {
-                        const videoRef = ref.current;
-                        const endTime = end || (videoRef && videoRef.duration);
-
-                        if (videoRef && videoRef.currentTime === endTime) {
-                            videoRef.currentTime = start;
-                            videoRef.play().catch(() => setHasVideoFallback(true));
-                        }
-                    },
-                    {passive: true},
-                );
-            }
-
             if (playVideo) {
                 ref.current.play().catch(() => setHasVideoFallback(true));
             }
@@ -94,7 +76,7 @@ const Video = (props: VideoAllProps) => {
                 className={b('react-player', videoClassName)}
                 src={src}
                 previewImgUrl={previewImg}
-                loop={Boolean(loop)}
+                loop={loop}
                 controls={controls}
                 muted={muted}
                 autoplay={autoplay && playVideo}